Lens.Micro.Platform

Description

This module is an approximation for Control.Lens from lens; by importing it you get all functions and instances from microlens, microlens-mtl, microlens-ghc, as well as the following instances:

at for HashMap

each and ix for

HashMap

Vector and variants (unboxed vectors, etc)

strict Text and lazy Text

_head, _tail, _init, _last for

Vector and variants

strict and lazy Text

strict and lazy for Text

Documentation

module Lens.Micro

module Lens.Micro.GHC

module Lens.Micro.Mtl

module Lens.Micro.TH

packed :: IsText t => Lens' String t

packed lets you convert between String and Text (strict or lazy). It can be used as a replacement for pack or as a way to modify some String if you have a function like Text -> Text.

unpacked :: IsText t => Lens' t String

unpacked is like packed but works in the opposite direction.

Orphan instances

Ixed Text

Instance details

Methods

ix :: Index Text -> Traversal' Text (IxValue Text)

Ixed Text

Instance details

Methods

ix :: Index Text -> Traversal' Text (IxValue Text)

Strict Text Text

Instance details

Methods

strict :: Lens' Text0 Text

lazy :: Lens' Text Text0

(a ~ Char, b ~ Char) => Each Text Text a b

Instance details

Methods

each :: Traversal Text Text a b

(a ~ Char, b ~ Char) => Each Text Text a b

Instance details

Methods

each :: Traversal Text Text a b

Cons Text Text Char Char

Instance details

Methods

_Cons :: Traversal Text Text (Char, Text) (Char, Text)

Cons Text Text Char Char

Instance details

Methods

_Cons :: Traversal Text Text (Char, Text) (Char, Text)

Snoc Text Text Char Char

Instance details

Methods

_Snoc :: Traversal Text Text (Text, Char) (Text, Char)

Snoc Text Text Char Char

Instance details

Methods

_Snoc :: Traversal Text Text (Text, Char) (Text, Char)

Unbox a => Ixed (Vector a)

Instance details

Methods

ix :: Index (Vector a) -> Traversal' (Vector a) (IxValue (Vector a))

Storable a => Ixed (Vector a)

Instance details

Methods

ix :: Index (Vector a) -> Traversal' (Vector a) (IxValue (Vector a))

Prim a => Ixed (Vector a)

Instance details

Methods

ix :: Index (Vector a) -> Traversal' (Vector a) (IxValue (Vector a))

Ixed (Vector a)

Instance details

Methods

ix :: Index (Vector a) -> Traversal' (Vector a) (IxValue (Vector a))

(Unbox a, Unbox b) => Each (Vector a) (Vector b) a b

Instance details

Methods

each :: Traversal (Vector a) (Vector b) a b

(Storable a, Storable b) => Each (Vector a) (Vector b) a b

Instance details

Methods

each :: Traversal (Vector a) (Vector b) a b

(Prim a, Prim b) => Each (Vector a) (Vector b) a b

Instance details

Methods

each :: Traversal (Vector a) (Vector b) a b

Each (Vector a) (Vector b) a b

Instance details

Methods

each :: Traversal (Vector a) (Vector b) a b

(Unbox a, Unbox b) => Cons (Vector a) (Vector b) a b

Instance details

Methods

_Cons :: Traversal (Vector a) (Vector b) (a, Vector a) (b, Vector b)

(Storable a, Storable b) => Cons (Vector a) (Vector b) a b

Instance details

Methods

_Cons :: Traversal (Vector a) (Vector b) (a, Vector a) (b, Vector b)

(Prim a, Prim b) => Cons (Vector a) (Vector b) a b

Instance details

Methods

_Cons :: Traversal (Vector a) (Vector b) (a, Vector a) (b, Vector b)

Cons (Vector a) (Vector b) a b

Instance details

Methods

_Cons :: Traversal (Vector a) (Vector b) (a, Vector a) (b, Vector b)

(Unbox a, Unbox b) => Snoc (Vector a) (Vector b) a b

Instance details

Methods

_Snoc :: Traversal (Vector a) (Vector b) (Vector a, a) (Vector b, b)

(Storable a, Storable b) => Snoc (Vector a) (Vector b) a b

Instance details

Methods

_Snoc :: Traversal (Vector a) (Vector b) (Vector a, a) (Vector b, b)

(Prim a, Prim b) => Snoc (Vector a) (Vector b) a b

Instance details

Methods

_Snoc :: Traversal (Vector a) (Vector b) (Vector a, a) (Vector b, b)

Snoc (Vector a) (Vector b) a b

Instance details

Methods

_Snoc :: Traversal (Vector a) (Vector b) (Vector a, a) (Vector b, b)

(Eq k, Hashable k) => Ixed (HashMap k a)

Instance details

Methods

ix :: Index (HashMap k a) -> Traversal' (HashMap k a) (IxValue (HashMap k a))

(Eq k, Hashable k) => At (HashMap k a)

Instance details

Methods

at :: Index (HashMap k a) -> Lens' (HashMap k a) (Maybe (IxValue (HashMap k a)))

c ~ d => Each (HashMap c a) (HashMap d b) a b

Instance details

Methods

each :: Traversal (HashMap c a) (HashMap d b) a b