Lens.Micro.Platform.Internal

Description

Documentation

class IsText t where

Methods

packed :: Lens' String t

packed lets you convert between String and Text (strict or lazy). It can be used as a replacement for pack or as a way to modify some String if you have a function like Text -> Text.

unpacked :: Lens' t String

unpacked is like packed but works in the opposite direction.